+++ utils/database.utils.ts
@@ -1,11 +1,17 @@
-import mysql from "mysql2";
+import { DataSource } from "typeorm";
+import ProductEntity from "../models/product.model";
+import Cart from "../models/cart.model";
 
-const pool = mysql.createPool({
+export const AppDataSource = new DataSource({
+  type: "mysql",
   host: process.env.DB_HOST,
-  port: parseInt(process.env.DB_PORT || "3306"),
-  user: process.env.DB_USER,
+  port: parseInt(process.env.DB_PORT || "3306", 10),
+  username: process.env.DB_USER,
   password: process.env.DB_PASSWORD,
   database: process.env.DB_NAME,
-}).promise();
+  entities: [ProductEntity, Cart],
+  synchronize: false,
+  logging: true,
+});
 
-export default pool;
+export default AppDataSource;